Enable disallow_subclassing_any
Disallows subclassing a value of type ``Any``.

This flag reports an error whenever a class subclasses a value of type ``Any``.
This may occur when the base class is imported from a module that doesn't
exist (when using :option:`--ignore-missing-imports`) or is ignored due to
:option:`--follow-imports=skip <--follow-imports>` or a ``# type: ignore``
comment on the ``import`` statement.

Since the module is silenced, the imported class is given a type of ``Any``. By
default mypy will assume that the subclass correctly inherited the base class
even though that may not actually be the case. This flag makes mypy raise an
error instead.