Add py.typed marker file (PEP 561) (#108)
* Add py.typed marker file (PEP 561)

* Remove mypy_path from mypy.ini

This is no longer needed because the installed package contains the py.typed
marker file (PEP 561), allowing mypy to access inline type information. If an
import is followed from a module located outside of the package, mypy no longer
needs MYPY_PATH to be able to extract type information from the package. The
only requirement is that our package must be installed into the environment mypy
is running it, for example the Nox session or the Poetry environment.

* Install the package into the Nox session for mypy