Fix tempfile usage in Poetry.export on Windows
Poetry.export uses tempfile.NamedTemporaryFile to create a temporary
requirements file, passing the name of the temporary file to `poetry export` via
its `--output` option.

Unfortunately, this does not work on Windows. On Windows NT and later, the name
attribute of the file-like object returned by tempfile.NamedTemporaryFile cannot
be used to open the file a second time, while the named temporary file is still
open.

https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile

Instead, use tempfile.TemporaryDirectory to create a temporary directory, and
store a file named `requirements.txt` in that directory. This technique should
be both secure and portable.