Enable end-of-line normalization with Git attributes
This setting converts line endings to UNIX-style (`LF`), whenever text
files are checked into and out of the repository.

Without this setting, the pre-commit hook for Prettier fails on the
Windows runner. GitHub Actions configures git with `core.autocrlf =
true`, because this is the default setting in Git for Windows.

* Enable end-of-line normalization with Git attributes

* Add note about end-of-line normalization to documentation

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@3086228