Run the Tests workflow on Windows and macOS (Cookiecutter)
On Windows, run `git add --renormalize .` after `git add .` to work
around an issue in Cookiecutter where files are always generated with
Windows-style line endings.

Also switch `$ErrorActionPreference` to `Continue` before staging the
files. The command completes with an exit status of 1, and produces the
warning "CRLF will be replaced by LF" for every file in the repository.

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@dc73f25