Add .cookiecutter.json to project
* Add .cookiecutter.json to project

This file records the JSON context used when generating the project. It
allows to upgrade the project with changes from the Cookiecutter, using
tools like [cupper].

[cupper]: https://github.com/senseyeio/cupper

* Add post generation hook to fix JSON indentation width

The jsonify extension distributed with Cookiecutter uses an indentation
width of four spaces. This conflicts with the default indentation width of
Prettier for JSON files. Prettier is run as a pre-commit hook in CI.

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@5189d63