Remove pre-commit hook for mypy
The repository-local hook is problematic because it depends on:

  (a) Poetry having been installed and added to the PATH
  (b) dependencies having been installed into the Poetry environment
  (c) other things? like being in the right directory

For example, pre-commit inside magit is currently broken for this
reason.

I'd rather not use the official pre-commit hook for mypy. Its README.md
says:

  By default, mypy will run with `mypy --ignore-missing-imports`,
  `pre-commit` runs `mypy` from an isolated virtualenv so it won't have
  access to those.

  https://github.com/pre-commit/mirrors-mypy

Note that we currently have the same problem, because the package is no
longer installed into the Nox session for mypy. That change may need to
be reverted.

Using Nox means we can easily support multiple Python versions. This is
also possible with pre-commit, but it leads to a lot of boilerplate.
(Which is why we no longer run Flake8 on multiple Python versions.)