Remove pre-commit hook for mypy
The repository-local hook is problematic because it means that pre-commit needs to be able to run Poetry with a valid environment, introducing complexity and hidden dependencies.

The [official pre-commit] hook for mypy uses `--ignore-missing-imports` because it runs `mypy` from an isolated virtualenv where imports for the dependencies are not accessible.

Running mypy from Nox means we can easily support multiple Python versions. This is
also possible with pre-commit, but leads to a lot of boilerplate.

[official pre-commit]: https://github.com/pre-commit/mirrors-mypy

* Remove pre-commit hook for mypy

* Remove documentation about mypy hook

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@ff24694