Fix missing underlines on some links
Links in the first paragraph of every section are not underlined. This is
because the CSS selector intended for badges matches these links
erroneously.

Precede the badges by an `.. rst-class:: badges` directive. This wraps them in
an HTML element with the badges class, which can easily be targeted in CSS.

  <p class="badges>
    ...
  </p>

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@d20a6a1