Only run the targeted Python version in each CI job
The mypy and tests jobs were running with any Python version installed
on the runner. If a registered Python version for the session was not
installed on the runner, they would print a message that the session was
skipped. Avoid all of this, and specifically run the session for the
targeted Python version (as specified in matrix.python-version).

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@b6e8d98