Remove Quickstart from README
The README on the GitHub repository page reflects the state of the master
branch, instead of the stable release. To reduce potential confusion, it's
better to keep the README short, and omit the Quickstart content. This content
is still available on Read the Docs, and now lives in quickstart.rst itself,
instead of being included from the README.

The README now consists of the badges, the logo, the usage, and the list of
features. Sphinx references are spread out over the README, the Quickstart
Guide, and the User Guide, with each of these including references from its
predecessors.

Original-Commit: cjolowicz/cookiecutter-hypermodern-python@ba28681