Fix Retrocookie workflow skipping CI-related changes
When `git diff-tree` is invoked with a path like
`.github/workflows/retrocookie.yml`, and the commit in question changes another
file under `.github`, the command reports `.github` as changed. This causes a
false positive in our skip logic, where we interpret the result as a change to
retrocookie.yml itself. As a result, changes to CI workflows were not pushed to
the project template.

The workflow now uses the `--name-only` option to `git diff-tree` instead of
`--quiet`. This allows us to grep for the excluded files in the git output
instead of relying on the exit code.