📚 [docs] Generate API documentation for package not `__main__`
`reference.rst` inserts the API documentation for the `__main__` module. That
makes little sense, as it only contains the Click entry point, whose docstring
is not intended for consumption by Sphinx.

Instead, insert API documentation for the package. This generates documentation
for anything in `__init__`. On project generation, this is only be the package
docstring, which is fine.