👷 Resolve conflicts in Retrocookie workflow for itself [skip ci]
When Dependabot updates actions/setup-python or actions/checkout, this resulted
in modifications to the retrocookie.yml workflow definition. As a consequence,
the Retrocookie workflow would fail because the commit can no longer be
imported, as the template does not have this file.

We can automate this common case by checking if retrocookie.yml had conflicts
and using `git rm` and `git cherry-pick --continue` to resolve them. Other
conflicts will still cause the workflow to fail.